1.amount from Double to BigDecimal 2.refactoring method createOffer

+++ src/main/java/pojos/deposit/Order.java
@@ -4,20 +4,22 @@ package pojos.deposit;
 import javax.annotation.Generated;
 import com.google.gson.annotations.Expose;
 
+import java.math.BigDecimal;
+
 @Generated("net.hexar.json2pojo")
 @SuppressWarnings("unused")
 public class Order {
 
     @Expose
-    private Double amount;
+    private BigDecimal amount;
     @Expose
     private String currency;
 
-    public Double getAmount() {
+    public BigDecimal getAmount() {
         return amount;
     }
 
-    public Order setAmount(Double amount) {
+    public Order setAmount(BigDecimal amount) {
         this.amount = amount;
         return this;
     }
